+++ src/Faker.Tests/CompanyTests.cs
@@ -203,5 +203,17 @@ namespace RimuTec.Faker.Tests {
          Assert.AreEqual(1, Regex.Matches(number, @"^[89]").Count);
          Assert.AreEqual(9, Regex.Matches(number, @"[0-9]").Count);
       }
+
+      [Test]
+      public void AustralianBusinessNumber_HappyDays() {
+         // arrange
+
+         // act
+         var number = Company.AustralianBusinessNumber();
+
+         // assert
+         Assert.IsFalse(string.IsNullOrWhiteSpace(number));
+         Assert.AreEqual(11, Regex.Matches(number, @"[0-9]").Count);
+      }
    }
 }