+++ releasenotes.md
@@ -2,6 +2,7 @@
 
 ## Version 0.10.0
 ### New Features
+- Company.AustralianBusinessNumber()
 - Company.CzechOrganizationNumber()
 - Company.FrenchSirenNumber()
 - Company.FrenchSiretNumber()